Add velyst_motiongfx crate
