+++ Cargo.toml
@@ -13,6 +13,7 @@ repository = "https://github.com/voxell-tech/motiongfx"
 motiongfx = { version = "0.3.0", path = "crates/motiongfx" }
 bevy_motiongfx = { version = "0.3.0", path = "crates/bevy_motiongfx" }
 peniko_motiongfx = { version = "0.3.0", path = "crates/peniko_motiongfx" }
+velyst_motiongfx = { version = "0.3.0", path = "crates/velyst_motiongfx" }
 
 # bevy
 bevy = { version = "0.18.1", default-features = false }
